Addresses double-promotion warnings in tests and demos directories

+++ tests/LIFTest/src/LIFTestProbe.hpp
@@ -40,8 +40,8 @@ private:
    double * stddevs;
    int * counts;
 
-   float endingTime; // The time, in the same units dt is in, at which to stop the test.
-   float tolerance; // Number of standard deviations that the observed rates can differ from the expected rates.
+   double endingTime; // The time, in the same units dt is in, at which to stop the test.
+   double tolerance; // Number of standard deviations that the observed rates can differ from the expected rates.
 };
